code/requirements.txt
spacy
tqdm
pandas
seaborn
spacy_language_detection
lingua-language-detector
scikit-learn
pymupdf
